feat(engine/ocaml<->rust): use one big `Types` module
We had two generated OCaml modules containing Rust types:
 - one for types defined in the frontend
 - one for the types defined in the rust engine

However, the rust engine uses many types from the
frontend, and that resulted in a lot of
deduplication of types, where we had to do manual
coercions.

This commit puts all types in the same OCaml
`Types` module. This is much nicer in terms of
deduplication: there is no deduplication.

However, there are some name clash that are solved
via a numbered suffix. To disambiguate, this
commit adds the module `rust_engine_types`.