feat(rengine): use `ExplicitDefId` instead of `DefId` for names
There was no good reason to only use `DefId`s.
`ExplicitDefId` adds an information compared to `DefId`s: is the name a constructor or not?

The names available in `crate::names` are now `ExplicitDefId`, not just `DefId`s.

This is very useful when constructing a expression
or a type using `crate::name::*`: there is now no
need to specify wether what we are building is a
constructor or not. This information is now
available directly.