feat(gh actions): job that creates a "this month in hax" skeleton
This adds a CI workflow running each month.
The workflow creates a skeleton to be filled for the blog post.
The skeleton contains a header, the list of PRs and contributor.

The action also push a draft PR, and tags the person that should be in charge of writing the blog post.
To find who should write the blog post, the script finds who wrote the N-1 last posts, and picks someone random from the N blog post authors not included in the N-1 last posts.
(`N` being the number of authors, who are hardcoded in the script)