fix(oengine): tuples of size 1 in F*
Wrongly, the F* backend assumes tuples of size 1 are equivalent to the underlying type, e.g. `(T,)` reduces to `T`.
While assuming that about tuples, a projection `tuple_of_size_1.0` (in Rust) will result in a projection in F* as well: `tuple_of_size_1._1`.
Since `tuple_of_size_1` is typed `T` and not `Tuple1 T` in F*, this is ill-typed.

This commit doesn't change the wrong translation of `(T,)` as `T`, but fixes the latter translation.

This commit fixes the THIR importer, which inserted 0 as tuple length in projectors.
This commit also makes the F* backend skip field projections on tuples of size one.

This is a quickfix; the F* backend in the OCaml engine being legacy, we don't want to maintain it further.