refactor(rengine): introduce FreshModule case for GlobalId
* Context: In preparation for a namespace mechanism in the Lean backend, we
needed a way to compare items identifiers with the identifier of the current
module (to strip prefixes). However, hax can introduce
[fresh_modules] (typically, for bundles), and they did not have a normal
identifier: as they were not present in the rust code, they do not get an
[ExplicitDefId] and only exist as metadata (the [moved] field of
[ConcreteId]). To be able to handle the identifier of normal rust modules as
well as fresh modules in a unified way, this commit adds a new case in the
[GlobalId] enum.

This commit also contains some refactoring.

* refactor: remove conversion function from [ConcreteId] to [GlobalId]
* feat: add a memoized conversion function from [TupleId] to [ConcreteId]